+++ sdks/js/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project are documented in this file.
 
+## [0.3.4] - 2026-08-26
+
+### Changed
+
+- Updated the JavaScript SDK release version and credential examples to the
+  `lz_` prefix.
+
 ## [0.2.6] - 2026-05-30
 
 ### Fixed